add 'expand' as third argument in shortcode_atts() so they can be filtered
shortcode_atts() merges the passed-in values and the defaults for the shortcode. It's third argument is $shortcode, which is supposed to be the slug of the shortcode. It's optional, but if it's missing then no one can filter the arguments with the "shortcode_atts_{$shortcode}" filter.

This patch just adds the third argument so I can set a bunch of defaults for the [expand] shortcode on my site.